refactor: Move package directory of discount into domain.
