+++ src/main/java/milkcoke/core/domain/discount/DiscountPolicy.java
@@ -1,4 +1,4 @@
-package milkcoke.core.discount;
+package milkcoke.core.domain.discount;
 
 import milkcoke.core.domain.member.Member;
