+++ src/main/java/milkcoke/core/domain/discount/FixDiscountPolicy.java
@@ -1,17 +1,19 @@
-package milkcoke.core.discount;
+package milkcoke.core.domain.discount;
 
 import milkcoke.core.domain.member.Grade;
 import milkcoke.core.domain.member.Member;
 
-public class RateDiscountPolicy implements DiscountPolicy {
+public class FixDiscountPolicy implements DiscountPolicy {
 
-    private final int discountPercent = 10;
+    private final int discountFixAmount = 1_000;
 
     @Override
     public int getDiscountPrice(Member member, int price) {
         if (member.getGrade() == Grade.VIP) {
-            return (int)(price * discountPercent * 0.01);
+            return discountFixAmount;
+        } else {
+            return 0;
         }
-        return 0;
     }
+
 }
