+++ src/main/resources/appConfig.xml
@@ -13,7 +13,7 @@
     </bean>
 
     <bean id="memberRepository" class="milkcoke.core.repository.member.MemoryMemberRepository" />
-    <bean id="discountPolicy" class="milkcoke.core.discount.RateDiscountPolicy" />
+    <bean id="discountPolicy" class="milkcoke.core.domain.discount.RateDiscountPolicy" />
 
 
 </beans>
