+++ src/test/java/milkcoke/core/discount/RateDiscountPolicyTest.java
@@ -1,5 +1,6 @@
 package milkcoke.core.discount;
 
+import milkcoke.core.domain.discount.RateDiscountPolicy;
 import milkcoke.core.domain.member.Grade;
 import milkcoke.core.domain.member.Member;
 import org.junit.jupiter.api.DisplayName;
